BugFix: restore 3D mapper to FreeBSD
The code that made the inclusion of the 3D mapper dependent on the OS being
Linux, Windows or macOs was mis-guided as the issue that was to be solved
was that Raspberry Pis do not have hardware that can support the ancient
*direct* dialect of OpenGL (pre-1.5) that we use.

{They support the OpenGLES form which is a reduced functionality OpenGL for
embedded devices but which started off at a level comparable to OpenGL 2.0}
These later versions only use the *indirect* form which is expressed in
OpenGL ShaderLanguage a.k.a. GLSL.}

It is erronous to depend on the OS being just those three because the
primary OS used with the Raspberry Pi - called Raspbian - is itself a
version of Linux - specifically Debian, which has been compiled with some
additional support for the Arm 7 hard-fp (with floating-point support in
the processor core hardware) and the external hardware that the units
posses.

This commit removes the OS dependent tests from the QMake and CMake project
files but leaves the handling of the WITH_3DMAPPER environmental variable
which needs to be set to "NO" for the Raspberry Pi and theoretically any
other system which does not have support for the old OpenGL used in the
3d Mapper display.

Also, it also adds in the two JSON related headers needed in ctelnet.cpp
that I have already found necessary in another PR. This is possibly because
I tend to build with another environmental variable ẀITH_UPDATER defined to
the value "NO" as well - although on the FreeBSD platform that is implicit.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>