BugFix: close IRC window when single profile session ends
This was reported on Discord by user Highope#9700 and I confirmed that when
an IRC session is opened for a single profile (NOT multi-playing) and then
the Mudlet application is closed the IRC session and the window are not
closed. Whilst the `/quit` command will terminate them a seg. fault
then immediately occurs because the `dlgIRC` destructor tries to save the
settings but needs to use the `mudlet::self()` pointer which is a dangling
one by that time.

This commit ensures that the IRC window is closed when the single profile
is closed along with the Mudlet application. If this is not required then
the change in the `mudlet.cpp` file could be omitted in which case the
other changes in the `dlgIRC.cpp` file prevent a segment fault from using
a `mudlet` pointer that is no longer valid.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>